requirements.txt
python-dotenv
aiogram==2.5
openai==0.28.1
langchain
langchain_community